handle undo/redo with tab and comments
but its buggy and doesn't fall back to the native events inbetween these ones
