+++ src/components/universe/index.ts
@@ -67,7 +67,7 @@ export class TextEditor extends ClassifiedElement {
 
   @property({ type: String, reflect: true }) id = 'myid'
   @property({ type: Boolean, attribute: 'wrap' }) wordWrap = false
-  @property() private text = `-- Creating a table with various data types and constraints
+  @property() text = `-- Creating a table with various data types and constraints
 CREATE TABLE employees (
     employee_id INT PRIMARY KEY, -- Primary key constraint
     first_name VARCHAR(50) NOT NULL, -- Not null constraint
@@ -129,10 +129,8 @@ SET temp_data = CONCAT(temp_data, ' - Updated with a long concatenation string t
   @state() highlightedCode?: DirectiveResult
   @state() private cache: Array<number> = []
   @state() lines: Array<TemplateResult> = []
-  @state() private undoStack: Array<{ text: string; selectionStart: number; selectionEnd: number }> = []
-  @state() private redoStack: Array<{ text: string; selectionStart: number; selectionEnd: number }> = []
 
-  private textareaRef: Ref<HTMLTextAreaElement> = createRef()
+  public textareaRef: Ref<HTMLTextAreaElement> = createRef()
   private displayedCodeRef: Ref<HTMLElement> = createRef()
   private lineNumbersRef: Ref<HTMLElement> = createRef()
   private scrollerRef: Ref<HTMLElement> = createRef()
@@ -145,7 +143,6 @@ SET temp_data = CONCAT(temp_data, ' - Updated with a long concatenation string t
     // Add resize event listeners
     window.addEventListener('resize', this.handleResize)
     this.addEventListener('resize', this.handleResize)
-    document.addEventListener('keydown', this.handleKeyboardEvents)
   }
 
   override disconnectedCallback() {
@@ -154,92 +151,15 @@ SET temp_data = CONCAT(temp_data, ' - Updated with a long concatenation string t
     // Remove resize event listeners
     window.removeEventListener('resize', this.handleResize)
     this.removeEventListener('resize', this.handleResize)
-    document.removeEventListener('keydown', this.handleKeyboardEvents)
   }
 
   private handleResize = () => {
     this.updateLineCache()
   }
 
-  private handleKeyboardEvents = (event: KeyboardEvent) => {
-    // REDO (Cmd+Shift+Z or Ctrl+Shift+Z)
-    if (event.key === 'z' && (event.metaKey || event.ctrlKey) && event.shiftKey) {
-      event.preventDefault()
-      this.handleRedo()
-    }
-    // UNDO (Cmd+Z or Ctrl+Z)
-    else if (event.key === 'z' && (event.metaKey || event.ctrlKey)) {
-      event.preventDefault()
-      this.handleUndo()
-    }
-    // REDO (Cmd+Y or Ctrl+Y)
-    else if (event.key === 'y' && (event.metaKey || event.ctrlKey)) {
-      event.preventDefault()
-      this.handleRedo()
-    }
-  }
-
   override firstUpdated(changedProperties: PropertyValueMap<this>) {
     super.firstUpdated(changedProperties)
     this.updateLineCache() // update again after initial render (otherwise line numbers may fail to handle text wrapping in the editor)
-  }
-
-  private saveStateToUndoStack() {
-    const textarea = this.textareaRef.value
-    if (textarea) {
-      this.undoStack.push({
-        text: this.text,
-        selectionStart: textarea.selectionStart,
-        selectionEnd: textarea.selectionEnd,
-      })
-      this.redoStack = [] // Clear redo stack on new action
-    }
-  }
-
-  private handleUndo() {
-    if (this.undoStack.length > 0) {
-      const previousState = this.undoStack.pop()
-      if (previousState) {
-        this.redoStack.push({
-          text: this.text,
-          selectionStart: this.textareaRef.value?.selectionStart || 0,
-          selectionEnd: this.textareaRef.value?.selectionEnd || 0,
-        })
-        this.text = previousState.text
-        const textarea = this.textareaRef.value
-        if (textarea) {
-          textarea.value = this.text
-          textarea.selectionStart = previousState.selectionStart
-          textarea.selectionEnd = previousState.selectionEnd
-        }
-        this.updateLineCache()
-      }
-    } else {
-      document.execCommand('undo')
-    }
-  }
-
-  private handleRedo() {
-    if (this.redoStack.length > 0) {
-      const nextState = this.redoStack.pop()
-      if (nextState) {
-        this.undoStack.push({
-          text: this.text,
-          selectionStart: this.textareaRef.value?.selectionStart || 0,
-          selectionEnd: this.textareaRef.value?.selectionEnd || 0,
-        })
-        this.text = nextState.text
-        const textarea = this.textareaRef.value
-        if (textarea) {
-          textarea.value = this.text
-          textarea.selectionStart = nextState.selectionStart
-          textarea.selectionEnd = nextState.selectionEnd
-        }
-        this.updateLineCache()
-      }
-    } else {
-      document.execCommand('redo')
-    }
   }
 
   override render() {
@@ -313,7 +233,7 @@ SET temp_data = CONCAT(temp_data, ' - Updated with a long concatenation string t
     this.updateLineCache()
   }
 
-  protected updateLineCache() {
+  public updateLineCache() {
     const previousValue = this.lines
     this.cache = []
 
@@ -379,8 +299,6 @@ SET temp_data = CONCAT(temp_data, ' - Updated with a long concatenation string t
   }
 
   public toggleLineComments() {
-    this.saveStateToUndoStack()
-
     const textarea = this.textareaRef.value
     const start = textarea!.selectionStart
     const end = textarea!.selectionEnd
@@ -456,8 +374,6 @@ SET temp_data = CONCAT(temp_data, ' - Updated with a long concatenation string t
   }
 
   public insertTabAtSelection() {
-    this.saveStateToUndoStack()
-
     const textarea = this.textareaRef.value
     const start = textarea!.selectionStart
     const end = textarea!.selectionEnd
